backend/requirements.txt
asgiref
requests
django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
transformers
nltk
textblob
newspaper3k
diffusers
tqdm
torch
transformers
pillow
accelerate
scipy
safetensors
openai
django-extensions


